backend/requirements.txt
flask>=2.3.2,<3.0.0
flask-cors>=3.0.10,<4.0.0
werkzeug>=2.3.7,<3.0.0
markupsafe>=2.1.1,<3.0.0
click>=8.1.3,<9.0.0
blinker>=1.6.2,<2.0.0

numpy>=1.25.0,<1.26.5
scipy>=1.10.1,<1.11.0
pillow>=10.0.0,<11.0.0
opencv-python-headless>=4.7.0.72,<5.0.0

torch>=2.8.0,<3.0.0
torchaudio>=2.8.0,<3.0.0
transformers>=4.31.0,<5.0.0
pytorchcv>=0.0.55

moviepy>=1.0.3,<2.0.0
librosa>=0.11.0,<1.0.0
pydub>=0.25.1,<1.0.0

scikit-learn>=1.2.2,<2.0.0

requests>=2.32.5,<3.0.0
python-multipart>=0.0.20,<1.0.0

matplotlib>=3.9.1,<4.0.0
seaborn>=0.13.2,<1.0.0

cython>=0.29.37,<1.0.0
pybind11>=2.10.4,<3.0.0
meson>=1.9.1,<2.0.0


apple